indicator group refactor indicator count
