+++ src/Module/IndicatorGroup/Components/DataSources/Programs/index.js
@@ -7,30 +7,15 @@ import Error from "../../../../../Shared/Componets/Error/ErrorAPIResult";
 import {useSetRecoilState} from "recoil";
 import {useGetIndicatorProgramSource} from "../../../../../Utils/Hooks/DataSource";
 import _ from "lodash";
-import {indicatorGroupPrograms} from "../../../../../Store/IndicatorGroup";
-
-
-const query = {
-    programs: {
-        resource: 'programStages',
-        params: (({dataElementId})=>({
-            fields: [
-                'program[id,displayName]'
-            ],
-            filter: [
-                `programStageDataElements.dataElement.id:eq:${dataElementId}`
-            ]
-        }))
-    }
-}
+import {indicatorGroupProgramDataElements, indicatorGroupPrograms} from "../../../../../Store/IndicatorGroup";
 
 
 export default  function Programs({sources}){
 
+    const engine=useDataEngine()
 
     const updatePrograms=useSetRecoilState(indicatorGroupPrograms)
-
-    const engine=useDataEngine()
+    const updateProgramDataElements=useSetRecoilState(indicatorGroupProgramDataElements)
 
     const {loading, error, data}=useGetIndicatorProgramSource(sources,engine)
 
@@ -41,10 +26,11 @@ export default  function Programs({sources}){
         return <Error error={error} />
     }
 
-    console.log(sources)
+
+    updateProgramDataElements((prev)=>{return _.concat(prev,sources?.prgDtEl)})
+
 
     const res=_.concat([],data?.attr??[],data?.prgInd??[],data?.prgDtEl??[])
-
     const allProgram=_.uniqWith(res,_.isEqual)
 
     //updating count its used in the facts component
