+++ src/Shared/Componets/DisplayFormula/Index.js
@@ -5,14 +5,14 @@ import {
     getFinalWordFormula,
     getFormulaSources,
     getWordData
-} from "../../Utils/Functions/FormulaFunctions";
-import {dataTypes} from "../../Utils/Models";
+} from "../../../Utils/Functions/FormulaFunctions";
+import {dataTypes} from "../../../Utils/Models";
 import PropTypes from "prop-types";
-import Loader from "./Loaders/Loader";
-import Error from "./Error/ErrorAPIResult";
-import {useGetData} from "../../Utils/Hooks";
+import Loader from "../Loaders/Loader";
+import Error from "../Error/ErrorAPIResult";
+import {useGetData} from "../../../Utils/Hooks";
 import {useSetRecoilState} from "recoil";
-import {dataElementsStateDictionary, programIndicatorStateDictionary} from "../../Store";
+import {dataElementsStateDictionary, programIndicatorStateDictionary} from "../../../Store";
 
 export default function DisplayFormula(props){
     //props
