+++ src/Shared/Componets/RelatedIndicatorTable/index.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import {useDataQuery} from "@dhis2/app-runtime";
 import React, {useEffect} from 'react'
 import i18n from "@dhis2/d2-i18n";
-import DisplayFormula from "../DisplayFormula";
+import DisplayFormula from "../DisplayFormula/Index";
 import Loader from "../Loaders/Loader";
 import Error from "../Error/ErrorAPIResult";
